hw_12/pyproject.toml
[tool.poetry]
name = "hw-12"
version = "0.1.0"
description = ""
authors = ["TarasSabat <tarsab25@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlalchemy = "^2.0.30"
psycopg2 = "^2.9.9"
psycopg2-binary = "^2.9.9"
myapplication = "^0.1.0"
pydantic-settings = "^2.2.1"
alembic = "^1.13.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
